Ikke restart app ved feil i kall mot oppgave api
