+++ src/main/kotlin/no/nav/syfo/Bootstrap.kt
@@ -5,12 +5,10 @@ import com.fasterxml.jackson.databind.DeserializationFeature
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.SerializationFeature
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import com.fasterxml.jackson.module.kotlin.readValue
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.Job
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import net.logstash.logback.argument.StructuredArguments.fields
 import no.nav.syfo.application.ApplicationServer
@@ -23,19 +21,14 @@ import no.nav.syfo.clients.KafkaConsumers
 import no.nav.syfo.clients.KafkaProducers
 import no.nav.syfo.db.Database
 import no.nav.syfo.db.VaultCredentialService
-import no.nav.syfo.model.ManuellOppgave
-import no.nav.syfo.model.Merknad
 import no.nav.syfo.oppgave.service.OppgaveService
-import no.nav.syfo.persistering.handleReceivedMessage
+import no.nav.syfo.persistering.MottattSykmeldingService
 import no.nav.syfo.service.ManuellOppgaveService
-import no.nav.syfo.util.LoggingMeta
 import no.nav.syfo.util.TrackableException
 import no.nav.syfo.vault.RenewVaultService
-import org.apache.kafka.clients.consumer.KafkaConsumer
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.net.URL
-import java.time.Duration
 import java.util.concurrent.TimeUnit
 
 val objectMapper: ObjectMapper = ObjectMapper()
@@ -91,15 +84,18 @@ fun main() {
     applicationState.ready = true
 
     RenewVaultService(vaultCredentialService, applicationState).startRenewTasks()
-
-    launchListeners(
+    val mottattSykmeldingService = MottattSykmeldingService(
+        kafkaConsumers.kafkaConsumerManuellOppgave,
         applicationState,
-        env,
-        kafkaConsumers,
+        env.syfoSmManuellTopic,
         database,
         oppgaveService,
         manuellOppgaveService
     )
+
+    createListener(applicationState) {
+        mottattSykmeldingService.startConsumer()
+    }
 }
 
 fun createListener(applicationState: ApplicationState, action: suspend CoroutineScope.() -> Unit): Job =
@@ -116,61 +112,3 @@ fun createListener(applicationState: ApplicationState, action: suspend Coroutine
             applicationState.ready = false
         }
     }
-
-fun launchListeners(
-    applicationState: ApplicationState,
-    env: Environment,
-    kafkaConsumers: KafkaConsumers,
-    database: Database,
-    oppgaveService: OppgaveService,
-    manuellOppgaveService: ManuellOppgaveService
-) {
-    createListener(applicationState) {
-        val kafkaConsumerManuellOppgave = kafkaConsumers.kafkaConsumerManuellOppgave
-
-        kafkaConsumerManuellOppgave.subscribe(listOf(env.syfoSmManuellTopic))
-        blockingApplicationLogic(
-            applicationState,
-            kafkaConsumerManuellOppgave,
-            database,
-            oppgaveService,
-            manuellOppgaveService
-        )
-    }
-}
-
-suspend fun blockingApplicationLogic(
-    applicationState: ApplicationState,
-    kafkaConsumer: KafkaConsumer<String, String>,
-    database: Database,
-    oppgaveService: OppgaveService,
-    manuellOppgaveService: ManuellOppgaveService
-) {
-    while (applicationState.ready) {
-        kafkaConsumer.poll(Duration.ofMillis(0)).forEach { consumerRecord ->
-            val receivedManuellOppgave: ManuellOppgave = objectMapper.readValue(consumerRecord.value())
-            val loggingMeta = LoggingMeta(
-                mottakId = receivedManuellOppgave.receivedSykmelding.navLogId,
-                orgNr = receivedManuellOppgave.receivedSykmelding.legekontorOrgNr,
-                msgId = receivedManuellOppgave.receivedSykmelding.msgId,
-                sykmeldingId = receivedManuellOppgave.receivedSykmelding.sykmelding.id
-            )
-            val receivedManuellOppgaveMedMerknad = receivedManuellOppgave.copy(
-                receivedSykmelding = receivedManuellOppgave.receivedSykmelding.copy(
-                    merknader = listOf(
-                        Merknad(type = "UNDER_BEHANDLING", beskrivelse = "Sykmeldingen er til manuell behandling")
-                    )
-                )
-            )
-
-            handleReceivedMessage(
-                receivedManuellOppgaveMedMerknad,
-                loggingMeta,
-                database,
-                oppgaveService,
-                manuellOppgaveService
-            )
-        }
-        delay(100)
-    }
-}
