+++ src/main/kotlin/no/nav/syfo/oppgave/exceptions/OpprettOppgaveException.kt
@@ -1,0 +1,3 @@
+package no.nav.syfo.oppgave.exceptions
+
+class OpprettOppgaveException(msg: String) : Exception(msg)
